Updated the dbcontext scaffolding, and began work on the database interactions with the Library. A DbSetInterfacer will be used for the generic methods inside DatabaseStorageRepository, so it doesn't have to worry about which interfacer is doing the work. It mimics the FactoryManager in how it will be built.

+++ StoreManager/StoreManager.Library/Data/CustomerData.cs
@@ -15,14 +15,14 @@ namespace StoreManager.Library.Data
         public string LastName { get; set; }
         public string Email { get; set; }
         public string PhoneNumber { get; set; }
-        public long AddressId { get; set; }
+        public int AddressId { get; set; }
         public DateTime BirthDate { get; set; }
-        public long? DefaultStoreLocationId { get; set; }
-        public List<long> OrderIds { get; set; }
+        public int? DefaultStoreLocationId { get; set; }
+        public List<int> OrderIds { get; set; }
 
         public CustomerData() { }
 
-        public CustomerData(string firstName, string lastName, string email, string phoneNumber, long address, DateTime birthDate, long? defaultStoreLocationId, List<long> orderIds) {
+        public CustomerData(string firstName, string lastName, string email, string phoneNumber, int address, DateTime birthDate, int? defaultStoreLocationId, List<int> orderIds) {
             FirstName = firstName;
             LastName = lastName;
             Email = email;
