+++ StoreManager/StoreManager.Library/Data/StoreData.cs
@@ -5,14 +5,14 @@ namespace StoreManager.Library.Data
     public class StoreData : NamedData
     {
         public override string Name { get; set; }
-        public List<long> OperatingLocationIds { get; set; }
-        public List<long> CustomerIds { get; set; }
-        public List<long> OrderIds { get; set; }
-        public Dictionary<long, int> Inventory { get; set; }
+        public List<int> OperatingLocationIds { get; set; }
+        public List<int> CustomerIds { get; set; }
+        public List<int> OrderIds { get; set; }
+        public Dictionary<int, int> Inventory { get; set; }
 
         public StoreData() { }
 
-        public StoreData(string name, List<long> operatingLocations, List<long> customers, Dictionary<long, int> inventory) {
+        public StoreData(string name, List<int> operatingLocations, List<int> customers, Dictionary<int, int> inventory) {
             Name = name;
             OperatingLocationIds = operatingLocations;
             CustomerIds = customers;
